Running main/train_bed_grasp.py for cross validation.
Haven't gotten it running completely, but got some of the code fixed.

Removed all mention of os.environ setting GPU usage. On Triton I can specify a GPU using the environment variable, and the GPU is a LOT faster than using the CPU.

The good news is, I can get inside the data manager class and (inside ) call the data augmentation and, CRITICALLY, I am able to run  which runs through the YOLO network. I have most of the paths set up.

However I still need to get the cross validation set up. This is NOT for the actual bed-making results but first to see the benefit of pre-training. I have sketches of modifications in the configuration file but still need toget the train_network andother methods to acommodate that. I also need to better understand the paths and to understand when weights do or do not get loaded into the network.

So, the agenda is to figure out how to get the cross validation to work the way I want it.

Actually, before that, I can probably do the data augmentation by writing a separate script and just loading the data points in and visualizing the results.