i can run the success net on the cal data now
Results aren't that good either, but this may be worthwhile to do first before grasping since it should be easier

I replaced the old loss with the cross entropy, but still don't see any systematic decrease in test error (nor when we try adam).

See the checklist of stuff to do in my notebook, and visualize the predictions/images.