fixed major issue with breaking up rollouts
previously we were only using the first grasp point out of anything in the list. Instead we need to ensure we get all the grasp data points in a rollout. A rollout will have length K (or K+1 in my data if I am using the slow one with the initial configuration as the first item). Then the grasp points need to be K/2, and the success points need to also be K/2, so check the lengths and sizes of minibatches